Add filtering by genre
The list of genres is pulled from the database on each request.
For some reason, the `ORDER BY` clause is not working and the genres
come back in database order instead of alphabetically.

It looks like the initial query has the correct `ORDER BY` clause,
but when it executes after `.Commit()`, the `ORDER BY` clause has
disappeared.  I cannot find documentation on this topic.